events.html
The next page is the Events Page containing information in cards of our upcoming events. Each card contains a date, location , title of the event and informaton about the event there is also a button at the button of the card that the user can click on if they want to find out more, this button contains a hyperlink to the contact us page or the enquiry page.